// apps/ziel-demo/src/generated/resources.ts
export const taxonomyTermAri = ari(
  "TaxonomyTerm",
  s.object({ spaceId: s.string(), environmentId: s.string(), kind: s.string(), id: s.string(), locale: s.string() }),
);